Issue #11: Changing class Motor and renaming MotorProcessor as MotorController
