+++ src/kpi_rover_ecu/CMakeLists.txt
@@ -12,13 +12,15 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR}/include)
 set(SOURCE_FILES
     src/main.cpp
     src/TCPServerReceiver.cpp
-    src/motorsProcessor.cpp
+    src/motorsController.cpp
+    src/motor.cpp
 )
 
 set(HEADER_FILES
     include/TCPServerReceiver.h
     include/config.h 
-    include/motorsProcessor.h		
+    include/motorsController.h
+    include/motor.h		
 )
 
 add_executable(${PROJECT_NAME} ${SOURCE_FILES} ${HEADER_FILES})
