+++ src/kpi_rover_ecu/src/TCPServerReceiver.cpp
@@ -1,6 +1,6 @@
 #include "TCPServerReceiver.h"
 #include "config.h"
-#include "motorsProcessor.h"
+#include "motorsController.h"
 
 char* get_primary_ip() {
     /*
